+++ CHANGELOG.md
@@ -12,3 +12,13 @@
 **BREAKING CHANGES:**
 
 - Interface changes after adding support for iPadOS >=17.5
+
+## 1.0.0
+
+- Update resolving of root view controller
+- Swift Package Manager support
+
+**BREAKING CHANGES:**
+
+- Support SDK 3.10+
+- Support Flutter 3.38.0+